Added error page views, organised user profile page routing.
Also amended the MembershipUser User class. It now works properly.
Anon users cannot access anything in MembersOnly folder.
If you navigate to a user profile page that the user doesn't exist for, you will be redirected to a 404 page. The conventions are broken for this and yet to be made clearer.